Telemetry reading refactor
